Refactor return to admin authentication logic in Drawer and Navbar components
